Doing stacking for all the past model

+++ test_titanic.py
@@ -4,9 +4,11 @@ import numpy as np
 import main
 import glob
 import shutil
+import os
 
 def test_titanic(test_dir="test_history"):
-    shutil.rmtree(test_dir)
+    if os.path.exists(test_dir):
+        shutil.rmtree(test_dir)
     TRAIN = 600
     # load data
     train = pd.read_csv("titanic.csv")
@@ -23,9 +25,9 @@ def test_titanic(test_dir="test_history"):
         return -accuracy_score(y_true, y_pred)
     
     num_trial = 5
-    he = HyperEnsamble(train[:TRAIN],target[:TRAIN], accuracy_error_func,None, test_dir, trial=num_trial)
+    he = HyperEnsamble(train,target[:TRAIN], accuracy_error_func,None, test_dir, trial=num_trial)
     he.find_best_models()
-    assert he.find_best_ensamble() < -0.76
+    assert he.find_best_session_ensamble() < -0.76
     validation_error = accuracy_error_func(target[TRAIN:],he.predict(train[TRAIN:]))
     print("final validation:",validation_error)
     assert validation_error < -0.76
@@ -35,17 +37,21 @@ def test_titanic(test_dir="test_history"):
     num_cached_model = len(he.mh.get_all_model()) 
 
     # reusing the result
-    he = HyperEnsamble(train[:TRAIN], target[:TRAIN], accuracy_error_func, None, test_dir, trial=1, model_confs = [main.SvmConf])
+    he = HyperEnsamble(train, target[:TRAIN], accuracy_error_func, None, test_dir, trial=1, model_confs = [main.SvmConf])
     he.find_best_models()
     assert len(he.mh.get_all_model()) == num_cached_model 
     assert len(list(glob.glob(he.mh.path_feature + "/*"))) == 1
 
     # need to train because feature changed
     train = train.drop(["Fare"], axis=1)
-    he = HyperEnsamble(train[:TRAIN], target[:TRAIN], accuracy_error_func, None, test_dir, trial=1, model_confs = [main.SvmConf])
+    he = HyperEnsamble(train, target[:TRAIN], accuracy_error_func, None, test_dir, trial=1, model_confs = [main.SvmConf])
     he.find_best_models()
     assert len(he.mh.get_all_model()) == num_cached_model + 1
     assert len(list(glob.glob(he.mh.path_feature + "/*"))) == 2
 
+    prediction = he.find_bigensamble(10)
+    validation_error = accuracy_error_func(target[TRAIN:], prediction)
+    assert validation_error < -0.76
+
 if __name__=="__main__":
-    test_titanic()
+    test_titanic(test_dir="test_directory")
